requirements.txt
pytest
kafka-python
pydantic
pytest-html
pytest-rerunfailures
